+++ packages/perseus-core/src/index.ts
@@ -215,6 +215,7 @@ export {
     getSorterPublicWidgetOptions,
     shuffleSorter,
     SORTER_MAX_HORIZONTAL_CARDS,
+    SORTER_MAX_CARDS,
 } from "./widgets/sorter/sorter-util";
 /** @hidden */
 export type {BlankPublicWidgetOptions} from "./widgets/blank/blank-util";