+++ packages/perseus-core/src/widgets/sorter/sorter-util.ts
@@ -2,6 +2,8 @@ import {constrainedShuffle, type RNG, seededRNG} from "../../utils/random-util";
 
 import type {PerseusSorterWidgetOptions} from "../../data-schema";
 
+export const SORTER_MAX_CARDS = 10;
+
 /**
  * Maximum number of cards Sorter to supports in horizontal layout.
  * If this number is exceeded, the cards render vertically even if