+++ packages/perseus-editor/src/editor-page.tsx
@@ -1,4 +1,4 @@
-import {components, ClassNames, Dependencies} from "@khanacademy/perseus";
+import {ClassNames, components, Dependencies} from "@khanacademy/perseus";
 import {
     getDefaultAnswerArea,
     parseAndMigratePerseusItem,
@@ -10,7 +10,6 @@ import _ from "underscore";
 import {A11yContext, createA11yContextValue} from "./components/a11y-context";
 import IssuesPanel from "./components/issues-panel";
 import JsonEditor from "./components/json-editor";
-import ViewportResizer from "./components/viewport-resizer";
 import ItemEditor from "./item-editor";
 import {createDeviceApiOptionsDeriver} from "./util/derive-device-api-options";
 import {gatherLinterIssues} from "./util/gather-linter-issues";
@@ -33,13 +32,6 @@ import type {
 
 const {HUD} = components;
 
-type OnChangeParams = {
-    jsonMode?: boolean;
-    question?: PerseusRenderer;
-    hints?: Hint[];
-    answerArea?: PerseusAnswerArea | null | undefined;
-};
-
 type Props = {
     /** Additional templates that the host application would like to display
      * within the Perseus Editor.
@@ -48,8 +40,9 @@ type Props = {
     apiOptions?: APIOptions;
     answerArea: PerseusAnswerArea; // related to the question,
     dependencies: PerseusDependenciesV2;
-    /** "Power user" mode. Shows the raw JSON of the question. */
-    developerMode: boolean;
+    // TODO(benchristel): remove developerMode after October 1, 2026
+    /** @deprecated - has no effect */
+    developerMode?: boolean;
     hints: Hint[]; // related to the question,
     /** A function which takes a file object (guaranteed to be an image) and
      * a callback, then calls the callback with the url where the image
@@ -59,15 +52,13 @@ type Props = {
     imageUploader?: ImageUploader;
     /** The content ID of the AssessmentItem being edited. */
     itemId: string;
-    /** Whether the question is displaying as JSON or if it is
-     * showing the editor itself with the rendering
-     * Only used in the perseus demos. Consider removing.
-     */
+    /** Shows a textarea with the raw JSON instead of the editor GUI. */
     jsonMode: boolean;
     /** A function which is called with the new JSON blob of content. */
-    onChange: (changed: OnChangeParams) => void;
-    /** A function which is called when the preview device changes. */
-    onPreviewDeviceChange: (arg1: DeviceType) => unknown;
+    onChange: (changed: PerseusItem) => void;
+    // TODO(benchristel): remove onPreviewDeviceChange after October 1, 2026
+    /** @deprecated - has no effect, and is never called */
+    onPreviewDeviceChange?: (arg1: DeviceType) => unknown;
     previewDevice: DeviceType;
     /** A global control to expand/collapse all widget editors on a page. */
     widgetsAreOpen?: boolean;
@@ -87,7 +78,6 @@ type Props = {
 
 type DefaultProps = {
     answerArea: Props["answerArea"];
-    developerMode: Props["developerMode"];
     hints: Props["hints"];
     jsonMode: Props["jsonMode"];
     onChange: Props["onChange"];
@@ -113,7 +103,6 @@ class EditorPage extends React.Component<Props, State> {
 
     static defaultProps: DefaultProps = {
         answerArea: getDefaultAnswerArea(),
-        developerMode: false,
         hints: [],
         jsonMode: false,
         onChange: () => {},
@@ -229,19 +218,6 @@ class EditorPage extends React.Component<Props, State> {
         });
     }
 
-    toggleJsonMode: () => void = () => {
-        this.setState(
-            {
-                json: this.serialize(),
-            },
-            () => {
-                this.props.onChange({
-                    jsonMode: !this.props.jsonMode,
-                });
-            },
-        );
-    };
-
     getSaveWarnings(): any {
         return this.itemEditor.current?.getSaveWarnings();
     }
@@ -262,7 +238,7 @@ class EditorPage extends React.Component<Props, State> {
         return this.itemEditor.current.serialize();
     }
 
-    handleChange = (toChange: OnChangeParams) => {
+    handleChange = (toChange: Partial<PerseusItem>) => {
         const newProps = _(this.props).pick("question", "hints", "answerArea");
         _(newProps).extend(toChange);
         this.props.onChange(newProps);
@@ -293,8 +269,6 @@ class EditorPage extends React.Component<Props, State> {
             touch,
         });
 
-        const showEditor = !this.props.developerMode || !this.props.jsonMode;
-
         if (deviceBasedApiOptions.isMobile) {
             className += " " + ClassNames.MOBILE;
         }
@@ -312,33 +286,6 @@ class EditorPage extends React.Component<Props, State> {
                 >
                     <div id="perseus" className={className}>
                         <div style={{marginBlockEnd: 10}}>
-                            {this.props.developerMode && (
-                                <span>
-                                    <label>
-                                        {" "}
-                                        Developer JSON Mode:{" "}
-                                        <input
-                                            type="checkbox"
-                                            checked={this.props.jsonMode}
-                                            disabled={
-                                                this.props.apiOptions
-                                                    ?.editingDisabled
-                                            }
-                                            onChange={this.toggleJsonMode}
-                                        />
-                                    </label>{" "}
-                                </span>
-                            )}
-
-                            {!this.props.jsonMode && (
-                                <ViewportResizer
-                                    deviceType={this.props.previewDevice}
-                                    onViewportSizeChanged={
-                                        this.props.onPreviewDeviceChange
-                                    }
-                                />
-                            )}
-
                             {!this.props.jsonMode && (
                                 <HUD
                                     message="Style warnings"
@@ -352,7 +299,7 @@ class EditorPage extends React.Component<Props, State> {
                                 />
                             )}
                         </div>
-                        {this.props.developerMode && this.props.jsonMode && (
+                        {this.props.jsonMode ? (
                             <div>
                                 <JsonEditor
                                     multiLine={true}
@@ -362,22 +309,20 @@ class EditorPage extends React.Component<Props, State> {
                                     editingDisabled={editingDisabled}
                                 />
                             </div>
-                        )}
-
-                        <A11yContext.Provider
-                            value={createA11yContextValue({
-                                setIssueHighlight: this.setIssueHighlight,
-                                a11yScanningEnabled:
-                                    this.state.a11yScanningEnabled,
-                                setA11yScanningEnabled:
-                                    this.setA11yScanningEnabled,
-                                highlightInstanceIds:
-                                    this.state.highlightInstanceIds,
-                                onA11yReport: this.handleA11yReport,
-                                axeCoreIssues: this.state.axeCoreIssues,
-                            })}
-                        >
-                            {showEditor && (
+                        ) : (
+                            <A11yContext.Provider
+                                value={createA11yContextValue({
+                                    setIssueHighlight: this.setIssueHighlight,
+                                    a11yScanningEnabled:
+                                        this.state.a11yScanningEnabled,
+                                    setA11yScanningEnabled:
+                                        this.setA11yScanningEnabled,
+                                    highlightInstanceIds:
+                                        this.state.highlightInstanceIds,
+                                    onA11yReport: this.handleA11yReport,
+                                    axeCoreIssues: this.state.axeCoreIssues,
+                                })}
+                            >
                                 <div className="perseus-editor-table">
                                     <div className="perseus-editor-row">
                                         <div className="perseus-editor-left-cell">
@@ -387,9 +332,6 @@ class EditorPage extends React.Component<Props, State> {
                                         </div>
                                     </div>
                                 </div>
-                            )}
-
-                            {showEditor && (
                                 <ItemEditor
                                     ref={this.itemEditor}
                                     itemId={this.props.itemId}
@@ -408,8 +350,8 @@ class EditorPage extends React.Component<Props, State> {
                                     highlightLint={this.state.highlightLint}
                                     problemNum={this.props.problemNum}
                                 />
-                            )}
-                        </A11yContext.Provider>
+                            </A11yContext.Provider>
+                        )}
                     </div>
                 </ItemEditorContext.Provider>
             </Dependencies.DependenciesContext.Provider>