+++ packages/perseus-linter/src/rules/sorter-widget-warning.test.ts
@@ -1,7 +1,8 @@
+import {SORTER_MAX_CARDS} from "@khanacademy/perseus-core";
+
 import {expectWarning} from "../__tests__/test-utils";
 
 import sorterWidgetWarningRule, {
-    sorterMaxCards,
     sorterMaxIdealCards,
     sorterMaxHorizontalCards,
 } from "./sorter-widget-warning";
@@ -16,7 +17,7 @@ describe("sorter-widget-warning", () => {
                     "sorter 1": {
                         options: {
                             correct: Array.from(
-                                {length: sorterMaxCards + 1},
+                                {length: SORTER_MAX_CARDS + 1},
                                 (_, i) => i,
                             ),
                             layout: "vertical",
@@ -40,7 +41,7 @@ describe("sorter-widget-warning", () => {
                     "sorter 1": {
                         options: {
                             correct: Array.from(
-                                {length: sorterMaxCards + 1},
+                                {length: SORTER_MAX_CARDS + 1},
                                 (_, i) => i,
                             ),
                             layout: "horizontal",